Use Markdown headers for issue template sections
This replaces the simple bolded text with `####` (H4) headers, which
render in a similar size w/in GitHub but add consistent margin between
the header and the following content. (Otherwise, whether the next line
of text appeared directly below the header or not would depend on if
the contributor entered a newline between them or not, and that’s just
sloppy.)

It also adds a much-needed period to one of the descriptions.

Signed-off-by: Justin Reese <justin@justinreese.com>