Create absolute path alias
We are using the `module-alias` package to allow for absolute path
imports throughout our code base. The plan is to use a single `$root`
alias for the entire project.

This sets up that alias. I would normally add a linting rule as well,
but there doesn't seem to be a good way to enforce it.

Issue #6