Create .env template
We'll be using dotenv for basic system configuration.  This is truly
intended to be devops / dev style configuration of the project.

We should be careful to not rely on dotenv config for things such as
specifying channel streams and other tool-configurations that ultimately
should be exposed to the end user of the tv-kitchen.

Issue #6