Mock up ingestion through a yarn script
This commit adds FFmpeg as a prerequisite and mocks up a simple
FFmpeg-based stream of input to one of our available internal stream
addresses. It doesn’t add or involve an actual ingestion engine;
instead, it demonstrates the stitching exists to build one.

Note that this commit dangerously passes the input from `yarn ingest`
directly to the `ffmpeg` command, which is explicitly warned against in
Node’s `spawn()` documentation. We will sanitize this.

Affects #15

Signed-off-by: Justin Reese <justin@justinreese.com>