Move ESLint configuration magic to shared util
To ensure ESLint is configured with our `module-alias` settings, we
dynamically generate part of its configuration file by loading up
`package.json`, extracting the `module-alias` configuration, and
transforming it into ESLint's expected format.

Only that last bit is ESLint-specific; the rest will be useful for
other tools that need to be configured with our aliases, such as Jest.

This commit moves the shared tooling to a central location and
establishes the pattern for tool-specific configuration generation.

Signed-off-by: Justin Reese <justin@justinreese.com>