Add ability to configure internal streams
This adds two new configuration options to the `.env.template` for
configuring the network address and port ranges that users want TV
Kitchen to use for internal stream communication, as well as a utility
script for listing them all. It does not add any checks about whether
they’re open or not, yet.

It also adds several utilities for expanding the shorthand syntax of
the port range option, and copious tests for those utilities.

Affects #14

Signed-off-by: Justin Reese <justin@justinreese.com>