Add mpegts utils
Adding our first utility function, which converts a pts / dts to a more
universally meaningful unit (seconds). The number of seconds a pts / dts
represents depends on the basetime of the mpegts stream it came from.
The default basetime is 90,000, which is why we use that default in the
utility.

Issue #45