Add ts-demuxer package
ffmpeg produces an mpegts stream, but in order to know exactly where in
the underlying video a given piece of that stream represents it needs to
be parsed.

Parsing an mpegts stream is not an art, but it does involve a fair bit
of code.  Luckily for us the `ts-demuxer` package exists.  This package
will be used by the AbstractIngestionEngine to process time from the
video stream before sending the data to kafka.

Issue #45