Add test data loader utility
This small utility will load json test data from the data directory. As
part of the utility I've also added some instructions to our TEST
documentation providing guidance on when to use a json file and how to
name it during a test.

The data loader only supports JSON, though in future we may want to
extend the method to intelligently parse json vs js data files and
either parse JSON or export the module default of a test data file.