Add time to AbstractIngestionEngine
This adds the ability to track time in a video stream before piping it
to kafka, decorating our payloads with time information.

This is a fairly significant shift to the way AbstractIngestionEngine
functions.

First, we have removed the pattern of directly passing
parameters to ffmpeg. In parallel, we have codified a required pattern
of piping a stream of data to ffmpeg over stdio.  This is a safer place
to start, and gives us more control over the speed at which data is
passed to ffmpeg.

Second, instead of passing data directly from ffmpeg to kafka we are
first processing it using the TSDemuxer tool in order to extract the
presentation timestamp (PTS).

Third, instead of using 'data' events we have moved to stream pipelines
for moving data around. Prior to this approach we were using pause
/ resume to prevent race conditions where the stream would be processed
faster than it was being written to kafka.  Pipelines have that control
built in.

This also adds a few tests to make the AbstractIngestionEngine test
suite more complete.

Many thanks to Tracey Jaquith, Chris Zubak-Skees, Matt Hirsch, and
Justin Reese for the R&D and discussion that led to this solution for
extracting time.

Issue #45 Track stream time

Co-authored-by: Chris Zubak-skees <chriszs@gmail.com>
Co-authored-by: Justin Reese <justin@justinreese.com>