Rename /src/lib to /src/tools
`lib` was an overloaded directory name for us, since we have
non-transpiled code in a `lib` directory that is a sibling of `src`.
We were using `src/lib` as a directory for various tools such as utils,
loggers, and 3rd party connection managers (e.g. kafka).

By renaming that directory to `tools` we keep the meaning without
causing potential confusion from having both `/lib` and `/src/lib`

Issue #46