Add countertop topology generation
This adds everything required to generate countertop topologies, where
a toplogy represents the network of data streams that make up the actual
byproducts of the TV Kitchen.

This commit introduces some core concepts:

1. CountertopCoordinator, which is the primary entry point of the
   countertop for developers who are creating a TV Kitchen instance.

2. CountertopStation, which houses a single type of configured
   appliance. A given station might have more than one active copy of
   that appliance, in order to support more than one distinct data
   stream.

3. CountertopStream, which is a thread of data that is being processed
   and tranformed by various appliances / stations.

There will be additional countertop concepts (e.g. the worker) but the
scope of this commit is simply to allow someone to register appliances
with the countertop and have the countertop generate the toplogy of
those appliances (streams and stations).

Topologies are basically a graph where vertices are stations, and edges
form paths which are captured by Streams. We store one stream per
cumulative path (so A, and A => B, A => B => C would each get a distinct
stream).

See issue #23 for a discussion on various topology examples that this
algorithm would generate and support.

Issue #23
Issue #30
Issue #82
Issue #6