Remove module-alias package
It was never really meant to be, was it.

We were using `module-alias` to avoid "relative import hell".
Specifically we used `%src` as an alias for the project root, allowing
us to import things as absolute paths from the base directory.

Unfortunately, module-alias is not a kind thing to include in npm
packages. Alternatives (e.g. link-module-alias) have the same problem.
now that this repository is being converted into a package instead of
a standalone application, we have to accept our fate and use relative
imports.

This removes the package, removes the shims, and updates paths to point
to the right place.

Issue #89