Add logger injection
Countertop logging is now injected into a given instance.  This will
make it possible to log using `this.logger` from within the Countertop
or CountertopStation.

This assumes that some of the more "Data Structure" (and utility)
classes will not actually invoke logging, which is appropriate for
a package architecture.

Issue #91