Remove old kafka
These kafka objects are now obsolete, and are not used by the
Countertop.

This takes out the old kafka, the old environment variable setting, and
also removes some of the utility scripts for spinning up a local kafka
instance.

It keeps the travis support for kafka, although it is still not being
used.  This is because integration tests will require it and the
repository was supporting travis kafka before.

Issue #99