Fix message ordering
CoutnertopWorker was writing to Kafka as quickly as it received events,
but there was nothing in place to ensure that Kafka was processing
a given message before the next one was sent.

This adds a Writable which will ensure Kafka has processed a given
payload before continuing to process the stream.

Issue #110