Add a default Kafka broker
KafkaJS requires a broker array to be passed to it, which means that we
were requiring users to pass kafkaSettings to our countertops by
default.

This was lame, and noisy, and felt way less magical than just being able
to say `new Countertop()`.

This adds a reasonable default (localhost with the default kafka port).

Issue #113