Allow incomplete tributaries
We had a condition that a new station can only be added to a topology if
it had at least one stream per input.  This restriction meant that
appliances could not exist with optional input types.

By removing this restriction we allow for optional inputs at the expense
of also potentially allowing a user to create a misconfigured topology
without receiving an error.

We may want to eventually assert that at least ONE input type is
populated for non-source appliances, but for now I didn't want to add
that restriction since this is still experimental and we may think of
a situation where appliances with no configured input provide some type
of value.

Issue #123