Set Kafka retention to 30 seconds
TV Kitchen generates a lot of data, and much of it is heavy stuff like
video and audio. Kafka, by default, will store data on its queue for
7 days. This means TV Kitchen will quickly run out of space when
processing streams of video.

This sets a 30 second retention limit for TV Kitchen topics. Data is
processed as it is generated so this limit should have no impact on the
functionality of the countertop.

Issue #133