Provide settings to appliance static methods
The getInputType and getOutputType methods of an appliance are used to
build topology maps.  By providing the related settings for a given
appliance we make it possible for those static methods to provide more
specific types, depending on the appliance's configuration.

This is related to an issue in the appliance repository:

https://github.com/tvkitchen/appliances/issues/151