Support TypeScript
This does not convert anything over to TypeScript in this repository but
it does allow us to start developing TypeScript and moving things in
pieces as appropriate.

This sets up all of our parts to be able to handle TypeScript and also
updates some of the versions of our dev tooling to support things like
TS linting.

Issue #153