Prune obsolete streams
When toplogies are created we want to be sure to prune any streams that
are just "less populated" versions of streams that were newly created
with the addition of a new station.

We also want to get rid of any streams that relied on those pruned
streams for input at any point in their tributary tree.

Issue #159