Narrow scope of export sorting
We don't actually want exports to always be locked into a specific
order; we DO want it for index files (which are just exporting all items
in a directory).  This changes the rule to only apply to those files,
giving us more flexibility to organize code in a way that makes sense in
other multi-export circumstances.