Fix fake timers
I apparently was not using fake timers correctly in any way shape or
form; I didn't realize this because jest runs in silent mode (I should
consider the wisdom of that setting...).

This uses fake timers properly; it also ensures that the timers are
actually being honored by validating the payload timestamps directly.