Use Ajv for PayloadParameter validation
This is our first use of Ajv to define a TypeGuard; it seems to work
well.

This might seem a bit redundant but I think that's just the nature of
the fact that TypeScript interfaces are not compiled to any sort of
runtime environment.

It also adds tests for the TypeGuard which should have been included
originally.