Add Avro support
Avro is a data serialization system[1].  We use it when reading and
writing Payloads to Kafka as part of the AvroPayload class.  This uses
a similar approach but instead of creating a whole new AvroPayload we
just add a new `byteSerialization` and `byteDeserialization` method to
the primary Payload class.

As part of this change I removed the "no extra fields" requirement from
they PayloadParameters schema / type guard.  It's possible this is the
wrong decision but I'm actually not used to type guards ensuring no
extra data (it seems like it might unnecessarily break backwards
compatibility in some cases), and it was causing a hurdle for me.

Avro also supports string serialization which is almost certainly faster
than JSON.stringify and JSON.parse.  That said: that form of
serialization isn't actually used anywhere at this point so I'm not
going to modify it yet.

Issue #154 Base should be baked into countertop

[1] https://avro.apache.org/