Add PayloadArray
This adds the PayloadArray class, which is a special data structure
which will keep an array of Payloads sorted in order of position.

There is a lot of history lost in this conversion, so I recommend
viewing the original commit history if you are confused.[1]

I am going to include this line, however, related to the sortedArray
utilities:

```
We are going to support a sorted payload array data type, and that will
involve a few methods that benefit from searches that are optimized
around sorted arrays. These methods do not need to be invented, but we
also don't want to bring in bloated dependencies.

After looking at the method-level lodash dependencies we found that the
best path would be to implement our own version of sortedIndexBy. The
lodash dependency represented 2,000 lines of code, whereas our
implementation is under 20.
```

Issue #154 Base should be baked into countertop

[1] https://github.com/tvkitchen/base/commits/main/packages/classes/src