Remove babel from build
It turns out we can have typescript process js files.  I'd like to start
doing that, which means we don't need to use babel for linting or
building.  It seems Jest still has a lot of trouble with ES6 module
syntax, so I'm keeping that for now even if it means having a few last
traces of babel in the repository.

My hope is that by having typescript process everything it will be
easier going forward to have js files import ts classes.

Issue #153