Configure babel to handle typescript
In the long run we want to eliminate babel entirely, but while the
migration to typescript is happening we'll want to be able to import ts
files from a js file.

This means updating babel to handle that situation; it also means
updating eslint to resolve ts when linting js files.

The typescript specific steps still exist (babel isn't generating type
files, and eslint needs to lint typescript files according to typescript
specific rules).

Issue #153